Model Validtaion Integer and AKO Decimal, Amount Decimal

+++ database/migrations/2026_03_12_134948_create_sales_models_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->text("detail");
             $table->date("date");
             $table->integer("purcahse_types_id")->comment('from purcahse types id');
-            $table->integer("quantity")->comment('');
+            $table->decimal("quantity",10,2)->comment('');
             $table->integer("financial_year_id");
             $table->timestamps();
         });
